# swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: api-gateway-ms
  description: Api gateway microservice
  version: 1.0.0
servers:
- url: /rest/api-gateway-api/v1.0/
paths:
  /get/validate-token-user:
    get:
      tags:
      - Zent Api Gateway
      summary: Valida el token del usuario
      description: |
        Api gateway que se encarga de validar el token del usuario para el aplicativo Zentinel tanto web  como movil
      operationId: get_validate_token_user
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
          example: Bearer fcea920f7412b5da7be0cf42b8c93759
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.api_gateway_controller.EventLiveView
components:
  schemas:
    GenericResponse:
      properties:
        data:
          type: object
          nullable: true
          example: Estructura generica.
        errorCode:
          type: integer
          example: 0
        message:
          type: string
          example: Mensaje generico.
        internalTransactionId:
          type: string
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
      example:
        errorCode: 0
        internalTransactionId: d290f1ee-6c54-4b01-90e6-d701748f0851
        externalTransactionId: fcea920f7412b5da7be0cf42b8c93759
        message: Mensaje generico.
        data: {}
    ResponseError:
      type: object
      properties:
        errorCode:
          type: integer
        internalTransactionId:
          type: string
        externalTransactionId:
          type: string
        message:
          type: string